+++ src/microservices/Microservices.IsIdentifiable/Reporting/Reports/FailureStoreReport.cs
@@ -1,11 +1,15 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Data;
+using System.IO;
 using System.Linq;
+using CsvHelper;
+using Microservices.IsIdentifiable.Failures;
 using Microservices.IsIdentifiable.Options;
 
 namespace Microservices.IsIdentifiable.Reporting.Reports
 {
-    internal class FailureStoreReport : FailureReport
+    public class FailureStoreReport : FailureReport
     {
         private readonly object _odtLock = new object();
         private readonly DataTable _dtAllFailures;
@@ -40,7 +44,7 @@ namespace Microservices.IsIdentifiable.Reporting.Reports
             base.AddDestinations(opts);
             Destinations.ForEach(d => d.WriteHeader((from dc in _dtAllFailures.Columns.Cast<DataColumn>() select dc.ColumnName).ToArray()));
         }
-
+        
         public override void Add(Failure failure)
         {
             lock (_odtLock)
@@ -65,6 +69,53 @@ namespace Microservices.IsIdentifiable.Reporting.Reports
         protected override void CloseReportBase()
         {
             Destinations.ForEach(d => d.WriteItems(_dtAllFailures));
+        }
+
+        public IEnumerable<Failure> Deserialize(FileInfo oldFile)
+        {
+            int lineNumber = 0;
+
+            using (var r = new CsvReader(new StreamReader(File.OpenRead(oldFile.FullName))))
+            {
+                if(r.Read())
+                    r.ReadHeader();
+                else
+                    yield break;
+                lineNumber ++;
+                // "Resource", "ResourcePrimaryKey", "ProblemField", "ProblemValue", "PartWords", "PartClassifications", "PartOffsets" 
+
+                while (r.Read())
+                {
+                    lineNumber++;
+                    var parts = new List<FailurePart>();
+
+                    try
+                    {
+
+                        var words = r["PartWords"].Split(Separator);
+                        var classes = r["PartClassifications"].Split(Separator);
+                        var offsets = r["PartOffsets"].Split(Separator);
+
+                        for(int i = 0 ; i < words.Length; i++)
+                            parts.Add(new FailurePart(words[i],
+                                (FailureClassification) Enum.Parse(typeof(FailureClassification), classes[i], true),
+                                int.Parse(offsets[i])));
+                    }
+                    catch (Exception e)
+                    {
+                        throw new Exception("Error Deserializing line " + lineNumber, e);
+                    }
+
+                    yield return new Failure( parts)
+                    {
+                        Resource = r["Resource"],
+                        ResourcePrimaryKey = r["ResourcePrimaryKey"],
+                        ProblemField = r["ProblemField"],
+                        ProblemValue = r["ProblemValue"],
+                    };
+
+                }
+            }
         }
     }
 }