+++ src/microservices/Microservices.IsIdentifiable/Runners/IsIdentifiableAbstractRunner.cs
@@ -40,9 +40,35 @@ namespace Microservices.IsIdentifiable.Runners
 
         /// <summary>
         /// Matches digits followed by a separator (: - \ etc) followed by more digits with optional AM / PM / GMT at the end
+        /// However this looks more like a time than a date and I would argue that times are not PII?
+        /// It's also not restrictive enough so matches too many non-PII numerics.
         /// </summary>
         readonly Regex _date = new Regex(
             @"\b\d+([:\-/\\]\d+)+\s?((AM)|(PM)|(GMT))?\b", RegexOptions.IgnoreCase);
+
+        // The following regex were adapted from:
+        // https://www.oreilly.com/library/view/regular-expressions-cookbook/9781449327453/ch04s04.html
+        // Separators are space slash dash
+
+        /// <summary>
+        /// Matches year last, i.e d/m/y or m/d/y
+        /// </summary>
+        readonly Regex _dateYearLast = new Regex(
+		    @"\b(?:(1[0-2]|0?[1-9])[ ]?[/-][ ]?(3[01]|[12][0-9]|0?[1-9])|(3[01]|[12][0-9]|0?[1-9])[ ]?[/-][ ]?(1[0-2]|0?[1-9]))[ ]?[/-][ ]?(?:[0-9]{2})?[0-9]{2}(\b|T)" // year last
+        );
+        /// <summary>
+        /// Matches year first, i.e y/m/d or y/d/m
+        /// </summary>
+        readonly Regex _dateYearFirst = new Regex(
+	    	@"\b(?:[0-9]{2})?[0-9]{2}[ ]?[/-][ ]?(?:(1[0-2]|0?[1-9])[ ]?[/-][ ]?(3[01]|[12][0-9]|0?[1-9])|(3[01]|[12][0-9]|0?[1-9])[ ]?[/-][ ]?(1[0-2]|0?[1-9]))(\b|T)" // year first
+        );
+        /// <summary>
+        /// Matches year missing, i.e d/m or m/d
+        /// </summary>
+        readonly Regex _dateYearMissing = new Regex(
+    		@"\b(?:(1[0-2]|0?[1-9])[ ]?[/-][ ]?(3[01]|[12][0-9]|0?[1-9])|(3[01]|[12][0-9]|0?[1-9])[ ]?[/-][ ]?(1[0-2]|0?[1-9]))(\b|T)" // year missing
+        );
+
 
         /// <summary>
         /// List of columns/tags which should not be processed.  This is automatically handled by the <see cref="Validate"/> method.
@@ -92,7 +118,16 @@ namespace Microservices.IsIdentifiable.Runners
             if (fi.Exists)
                 LoadRules(File.ReadAllText(fi.FullName));
             else
-                _logger.Info("No Rules Yaml file found (thats ok)");
+                _logger.Info("No default Rules Yaml file found (thats ok)");
+
+            if (!string.IsNullOrWhiteSpace(opts.RulesFile))
+            {
+                fi = new FileInfo(_opts.RulesFile);
+                if (fi.Exists)
+                    LoadRules(File.ReadAllText(fi.FullName));
+                else
+                    throw new Exception("Error reading "+_opts.RulesFile);
+            }
 
             IWhitelistSource source = null;
 
@@ -214,7 +249,14 @@ namespace Microservices.IsIdentifiable.Runners
 
             if (!_opts.IgnoreDatesInText)
             {
-                foreach (Match m in _date.Matches(fieldValue))
+                foreach (Match m in _dateYearFirst.Matches(fieldValue))
+                    yield return new FailurePart(m.Value.TrimEnd(), FailureClassification.Date, m.Index);
+
+                foreach (Match m in _dateYearLast.Matches(fieldValue))
+                    yield return new FailurePart(m.Value.TrimEnd(), FailureClassification.Date, m.Index);
+
+                // XXX this may cause a duplicate failure if one above yields
+                foreach (Match m in _dateYearMissing.Matches(fieldValue))
                     yield return new FailurePart(m.Value.TrimEnd(), FailureClassification.Date, m.Index);
 
                 foreach (Match m in _symbolThenMonth.Matches(fieldValue))